requirements.txt
asgiref==3.7.2
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.3.1
cryptography==41.0.3
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
drf-spectacular==0.26.4
exceptiongroup==1.1.3
factory-boy==3.3.0
Faker==19.6.1
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.1
pluggy==1.3.0
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pytest==7.4.2
pytest-cov==4.1.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.3
rsa==4.9
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
